use the kit_reader to read files, don't chdir
Changing to the kit dir before rendering is a bit of a dirty hack,
but it works... as long as the kit's kitreader is Dir.  If you're
using something more complicated, like KitReader::SWAK, it doesn't
actually help.

Instead of doing that, this commit adds a mixin to the micromason
object, converting its reads of files-by-name into reads of kit
contents.

Admittedly, the way that the renderer is shoved into the mason
object is also a dirty hack, but presumably it's one that can be
corrected by someone more familiar with how micromason mixins are
meant to work.